requirements.txt
# YOLO目标检测训练和可视化依赖

# 核心框架
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.23.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 图像处理
PyWavelets>=1.4.0
scipy>=1.10.0

# 其他工具
tqdm>=4.65.0
psutil>=5.9.0
